Move permissions methods out of HasRoles trait
