+++ src/Traits/HasRoles.php
@@ -5,7 +5,6 @@ namespace Spatie\Permission\Traits;
 use Illuminate\Support\Collection;
 use Spatie\Permission\Contracts\Role;
 use Illuminate\Database\Eloquent\Builder;
-use Spatie\Permission\Contracts\Permission;
 use Illuminate\Database\Eloquent\Relations\MorphToMany;
 
 trait HasRoles
@@ -20,7 +19,6 @@ trait HasRoles
             }
 
             $model->roles()->detach();
-            $model->permissions()->detach();
         });
     }
 
@@ -35,20 +33,6 @@ trait HasRoles
             config('permission.table_names.model_has_roles'),
             'model_id',
             'role_id'
-        );
-    }
-
-    /**
-     * A model may have multiple direct permissions.
-     */
-    public function permissions(): MorphToMany
-    {
-        return $this->morphToMany(
-            config('permission.models.permission'),
-            'model',
-            config('permission.table_names.model_has_permissions'),
-            'model_id',
-            'permission_id'
         );
     }
 
@@ -85,65 +69,6 @@ trait HasRoles
                 }
             });
         });
-    }
-
-    /**
-     * @param string|array|\Spatie\Permission\Contracts\Permission|\Illuminate\Support\Collection $permissions
-     *
-     * @return array
-     */
-    protected function convertToPermissionModels($permissions): array
-    {
-        if ($permissions instanceof Collection) {
-            $permissions = $permissions->toArray();
-        }
-
-        $permissions = array_wrap($permissions);
-
-        return array_map(function ($permission) {
-            if ($permission instanceof Permission) {
-                return $permission;
-            }
-
-            return app(Permission::class)->findByName($permission, $this->getDefaultGuardName());
-        }, $permissions);
-    }
-
-    /**
-     * Scope the model query to certain permissions only.
-     *
-     * @param \Illuminate\Database\Eloquent\Builder $query
-     * @param string|array|\Spatie\Permission\Contracts\Permission|\Illuminate\Support\Collection $permissions
-     *
-     * @return \Illuminate\Database\Eloquent\Builder
-     */
-    public function scopePermission(Builder $query, $permissions): Builder
-    {
-        $permissions = $this->convertToPermissionModels($permissions);
-
-        $rolesWithPermissions = array_unique(array_reduce($permissions, function ($result, $permission) {
-            return array_merge($result, $permission->roles->all());
-        }, []));
-
-        return $query->
-            where(function ($query) use ($permissions, $rolesWithPermissions) {
-                $query->whereHas('permissions', function ($query) use ($permissions) {
-                    $query->where(function ($query) use ($permissions) {
-                        foreach ($permissions as $permission) {
-                            $query->orWhere(config('permission.table_names.permissions').'.id', $permission->id);
-                        }
-                    });
-                });
-                if (count($rolesWithPermissions) > 0) {
-                    $query->orWhereHas('roles', function ($query) use ($rolesWithPermissions) {
-                        $query->where(function ($query) use ($rolesWithPermissions) {
-                            foreach ($rolesWithPermissions as $role) {
-                                $query->orWhere(config('permission.table_names.roles').'.id', $role->id);
-                            }
-                        });
-                    });
-                }
-            });
     }
 
     /**
@@ -271,107 +196,11 @@ trait HasRoles
     }
 
     /**
-     * Determine if the model may perform the given permission.
-     *
-     * @param string|\Spatie\Permission\Contracts\Permission $permission
-     * @param string|null $guardName
-     *
-     * @return bool
-     */
-    public function hasPermissionTo($permission, $guardName = null): bool
-    {
-        if (is_string($permission)) {
-            $permission = app(Permission::class)->findByName(
-                $permission,
-                $guardName ?? $this->getDefaultGuardName()
-            );
-        }
-
-        return $this->hasDirectPermission($permission) || $this->hasPermissionViaRole($permission);
-    }
-
-    /**
-     * Determine if the model has any of the given permissions.
-     *
-     * @param array ...$permissions
-     *
-     * @return bool
-     */
-    public function hasAnyPermission(...$permissions): bool
-    {
-        if (is_array($permissions[0])) {
-            $permissions = $permissions[0];
-        }
-
-        foreach ($permissions as $permission) {
-            if ($this->hasPermissionTo($permission)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    /**
-     * Determine if the model has, via roles, the given permission.
-     *
-     * @param \Spatie\Permission\Contracts\Permission $permission
-     *
-     * @return bool
-     */
-    protected function hasPermissionViaRole(Permission $permission): bool
-    {
-        return $this->hasRole($permission->roles);
-    }
-
-    /**
-     * Determine if the model has the given permission.
-     *
-     * @param string|\Spatie\Permission\Contracts\Permission $permission
-     *
-     * @return bool
-     */
-    public function hasDirectPermission($permission): bool
-    {
-        if (is_string($permission)) {
-            $permission = app(Permission::class)->findByName($permission, $this->getDefaultGuardName());
-
-            if (! $permission) {
-                return false;
-            }
-        }
-
-        return $this->permissions->contains('id', $permission->id);
-    }
-
-    /**
      * Return all permissions the directory coupled to the model.
      */
     public function getDirectPermissions(): Collection
     {
         return $this->permissions;
-    }
-
-    /**
-     * Return all the permissions the model has via roles.
-     */
-    public function getPermissionsViaRoles(): Collection
-    {
-        return $this->load('roles', 'roles.permissions')
-            ->roles->flatMap(function ($role) {
-                return $role->permissions;
-            })->sort()->values();
-    }
-
-    /**
-     * Return all the permissions the model has, both directly and via roles.
-     */
-    public function getAllPermissions(): Collection
-    {
-        return $this->permissions
-            ->merge($this->getPermissionsViaRoles())
-            ->sort()
-            ->values();
     }
 
     public function getRoleNames(): Collection
